# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a golyadmin Server",
        "title": "Swagger API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
...
        "/system/role/menu": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "获取角色全部列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取角色全部列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/role/menulist": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "分页获取角色列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分页获取角色列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/role/role": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "新增角色",
                "parameters": [
                    {
                        "description": "LyadminRole模型",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.LyadminRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "新增角色",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "根据ID删除角色",
                "parameters": [
                    {
                        "description": "ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除角色",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/role/role/:id": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "编辑角色",
                "parameters": [
                    {
                        "description": "分页",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.LyadminRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "编辑菜单",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.Id": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer"
                }
            }
        },
...
        "response.StructResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "system.LyadminDept": {
            "type": "object",
            "properties": {
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                }
            }
        },
        "system.LyadminMenu": {
            "type": "object",
            "properties": {
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "component": {
                    "description": "对应前端文件路径",
                    "type": "string"
                },
                "component_name": {
                    "description": "对应前端文件名称",
                    "type": "string"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "is_catalog": {
                    "description": "是否目录",
                    "type": "boolean"
                },
                "is_link": {
                    "description": "是否外链",
                    "type": "boolean"
                },
                "keep_alive": {
                    "description": "是否缓存页面",
                    "type": "boolean"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "sort": {
                    "description": "显示顺序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态",
                    "type": "boolean"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                },
                "visible": {
                    "description": "是否显示菜单",
                    "type": "boolean"
                },
                "web_path": {
                    "description": "路由地址",
                    "type": "string"
                }
            }
        },
        "system.LyadminMenuButton": {
            "type": "object",
            "properties": {
                "api": {
                    "type": "string"
                },
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "menu_id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "system.LyadminRole": {
            "type": "object",
            "properties": {
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "dept": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.LyadminDept"
                    }
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "menu": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.LyadminMenu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "permission": {
                    "description": "manytomany",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.LyadminMenuButton"
                    }
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}